MainApp/requirements.txt
click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
regex
mysql-connector-python
mysql.connector
mysql_connector
mysql-connector